@@ -16,113 +16,111 @@
   },
   "private": true,
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.1.1",
-    "@angular-devkit/core": "16.1.1",
-    "@angular-devkit/schematics": "16.1.1",
-    "@angular-eslint/eslint-plugin": "16.1.0",
-    "@angular-eslint/eslint-plugin-template": "16.1.0",
-    "@angular-eslint/template-parser": "16.1.0",
-    "@angular/cli": "~16.1.0",
-    "@angular/compiler-cli": "16.1.2",
-    "@angular/language-service": "16.1.2",
-    "@commitlint/cli": "^17.6.3",
-    "@commitlint/config-conventional": "^17.6.3",
-    "@commitlint/config-nx-scopes": "^17.6.4",
+    "@angular-devkit/build-angular": "17.0.0",
+    "@angular-devkit/core": "17.0.0",
+    "@angular-devkit/schematics": "17.0.0",
+    "@angular-eslint/eslint-plugin": "17.2.0",
+    "@angular-eslint/eslint-plugin-template": "17.2.0",
+    "@angular-eslint/template-parser": "17.2.0",
+    "@angular/cli": "~17.0.0",
+    "@angular/compiler-cli": "17.0.2",
+    "@angular/language-service": "17.0.2",
+    "@commitlint/cli": "^18.2.0",
+    "@commitlint/config-conventional": "^18.1.0",
+    "@commitlint/config-nx-scopes": "^18.3.0",
     "@golevelup/ts-jest": "^0.3.5",
-    "@mands/nx-playwright": "^0.2.0",
-    "@nestjs/schematics": "10.0.1",
-    "@nestjs/testing": "10.0.3",
-    "@ngneat/spectator": "^15.0.0",
-    "@playwright/test": "^1.26.0",
-    "@schematics/angular": "16.1.1",
+    "@nestjs/schematics": "10.0.2",
+    "@nestjs/testing": "10.2.4",
+    "@ngneat/spectator": "^16.0.0",
+    "@nx/angular": "17.2.8",
+    "@nx/devkit": "17.2.8",
+    "@nx/eslint": "17.2.8",
+    "@nx/eslint-plugin": "17.2.8",
+    "@nx/jest": "17.2.8",
+    "@nx/js": "17.2.8",
+    "@nx/nest": "17.2.8",
+    "@nx/node": "17.2.8",
+    "@nx/playwright": "17.2.8",
+    "@nx/web": "17.2.8",
+    "@nx/webpack": "17.2.8",
+    "@nx/workspace": "17.2.8",
+    "@playwright/test": "^1.39.0",
+    "@schematics/angular": "17.0.0",
     "@total-typescript/ts-reset": "^0.3.7",
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
+    "@types/dompurify": "^3.0.5",
     "@types/jest": "29.4.4",
-    "@types/node": "18.16.19",
-    "@typescript-eslint/eslint-plugin": "5.61.0",
-    "@typescript-eslint/parser": "5.61.0",
-    "autoprefixer": "^10.4.0",
+    "@types/node": "20.11.5",
+    "@types/trusted-types": "^2.0.7",
+    "@typescript-eslint/eslint-plugin": "6.14.0",
+    "@typescript-eslint/parser": "6.14.0",
     "cobertura-merge": "^1.0.4",
-    "copy-webpack-plugin": "^11.0.0",
-    "eslint": "8.45.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-playwright": "^0.22.0",
     "eslint-plugin-rxjs": "^5.0.2",
     "husky": "^8.0.0",
     "jest": "29.5.0",
     "jest-environment-jsdom": "29.5.0",
     "jest-environment-node": "^29.5.0",
-    "jest-preset-angular": "13.1.0",
-    "lcov-result-merger": "^4.0.0",
+    "jest-preset-angular": "13.1.4",
+    "lcov-result-merger": "^5.0.0",
     "mongo-migrate-ts": "^1.1.4",
-    "nx": "16.5.3",
-    "playwright": "^1.26.0",
-    "postcss": "^8.4.5",
-    "prettier": "2.8.8",
-    "prettier-plugin-tailwindcss": "^0.4.0",
-    "supertest": "^6.3.3",
-    "tailwindcss": "^3.0.2",
+    "nx": "17.2.8",
+    "playwright": "^1.39.0",
+    "prettier": "3.2.4",
     "ts-jest": "29.1.0",
-    "ts-node": "10.9.1",
-    "typescript": "5.1.3",
-    "wait-on": "^7.0.1",
-    "@nx/workspace": "16.5.3",
-    "@nx/js": "16.5.3",
-    "@nx/eslint-plugin": "16.5.3",
-    "@nx/jest": "16.5.3",
-    "@nx/linter": "16.5.3",
-    "@nx/node": "16.5.3",
-    "@nx/nest": "16.5.3",
-    "@nx/webpack": "16.5.3",
-    "@nx/angular": "16.5.3"
+    "ts-node": "10.9.2",
+    "typescript": "5.2.2",
+    "wait-on": "^7.0.1"
   },
   "dependencies": {
-    "@angular/animations": "16.1.2",
-    "@angular/common": "16.1.2",
-    "@angular/compiler": "16.1.2",
-    "@angular/core": "16.1.2",
-    "@angular/forms": "16.1.2",
-    "@angular/platform-browser": "16.1.2",
-    "@angular/platform-browser-dynamic": "16.1.2",
-    "@angular/router": "16.1.2",
-    "@apollo/client": "^3.7.17",
-    "@apollo/server": "^4.5.0",
+    "@angular/animations": "17.0.2",
+    "@angular/common": "17.0.2",
+    "@angular/compiler": "17.0.2",
+    "@angular/core": "17.0.2",
+    "@angular/forms": "17.0.2",
+    "@angular/platform-browser": "17.0.2",
+    "@angular/platform-browser-dynamic": "17.0.2",
+    "@angular/router": "17.0.2",
+    "@apollo/client": "^3.8.10",
+    "@apollo/server": "^4.10.0",
     "@automapper/classes": "^8.7.7",
     "@automapper/core": "^8.7.7",
     "@nestjs/apollo": "^12.0.3",
-    "@nestjs/axios": "^3.0.0",
-    "@nestjs/common": "10.0.3",
+    "@nestjs/axios": "^3.0.1",
+    "@nestjs/common": "10.2.4",
     "@nestjs/config": "^3.0.0",
-    "@nestjs/core": "10.0.3",
+    "@nestjs/core": "10.2.4",
     "@nestjs/cqrs": "^10.0.1",
     "@nestjs/graphql": "^12.0.3",
-    "@nestjs/mongoose": "^10.0.0",
-    "@nestjs/platform-express": "10.0.3",
-    "@opentelemetry/instrumentation-graphql": "^0.35.0",
-    "@opentelemetry/instrumentation-mongoose": "^0.33.0",
+    "@nestjs/mongoose": "^10.0.2",
+    "@nestjs/platform-express": "10.2.4",
+    "@opentelemetry/instrumentation-graphql": "^0.36.0",
+    "@opentelemetry/instrumentation-mongoose": "^0.34.0",
     "@opentelemetry/instrumentation-pino": "^0.34.0",
-    "@opentelemetry/sdk-node": "^0.41.0",
+    "@opentelemetry/sdk-node": "^0.47.0",
     "@opentelemetry/sdk-trace-base": "^1.14.0",
-    "@sentry/angular-ivy": "^7.57.0",
-    "@sentry/integrations": "^7.57.0",
-    "@sentry/node": "^7.52.1",
-    "@sentry/opentelemetry-node": "^7.52.1",
-    "@sentry/profiling-node": "^1.0.0",
+    "@sentry/angular-ivy": "^7.77.0",
+    "@sentry/node": "^7.77.0",
+    "@sentry/opentelemetry-node": "^7.77.0",
     "@timonmasberg/automapper-nestjs": "^0.0.0",
-    "angular-oauth2-oidc": "^15.0.1",
-    "apollo-angular": "^5.0.0",
+    "angular-oauth2-oidc": "^17.0.0",
+    "apollo-angular": "^6.0.0",
     "axios": "^1.0.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
+    "dompurify": "^3.0.8",
     "graphql": "^16.7.1",
     "graphql-sse": "^2.2.0",
-    "mongoose": "^7.0.3",
-    "pino": "^8.14.1",
-    "pino-pretty": "^10.0.0",
+    "mongoose": "^8.0.0",
+    "ng-zorro-antd": "^17.1.0",
+    "pino": "^8.15.0",
+    "pino-pretty": "^10.2.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "0.13.0"
+    "zone.js": "0.14.2"
   }
 }
