@@ -1,7 +1,12 @@
-import { HttpClientModule } from '@angular/common/http';
-import { NgModule } from '@angular/core';
+import { registerLocaleData } from '@angular/common';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import de from '@angular/common/locales/de';
+import { APP_INITIALIZER, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
+import DOMPurify from 'dompurify';
+import { NZ_I18N, de_DE } from 'ng-zorro-antd/i18n';
+import { firstValueFrom } from 'rxjs';
 
 import { AuthModule, DevAuthModule } from '@kordis/spa/auth';
 import { GraphqlModule } from '@kordis/spa/graphql';
@@ -15,17 +20,19 @@ import { AppComponent } from './component/app.component';
 import { ProtectedComponent } from './component/protected.component';
 import routes from './routes';
 
+registerLocaleData(de);
+
 @NgModule({
 	declarations: [AppComponent, ProtectedComponent],
 	imports: [
 		BrowserModule,
-		HttpClientModule,
 		RouterModule.forRoot(routes),
+		HttpClientModule,
 		environment.oauth
 			? AuthModule.forRoot(
 					environment.oauth.config,
 					environment.oauth.discoveryDocumentUrl,
-			  )
+				)
 			: DevAuthModule.forRoot(),
 		GraphqlModule.forRoot(
 			environment.apiUrl + '/graphql',
@@ -36,8 +43,40 @@ import routes from './routes';
 					environment.sentryKey,
 					environment.environmentName,
 					environment.releaseVersion,
-			  )
+				)
 			: NoopObservabilityModule.forRoot(),
+	],
+	providers: [
+		{
+			provide: APP_INITIALIZER,
+			// we need to use a global tt policy here mainly for the ant design icons
+			useFactory: () => () => {
+				globalThis.trustedTypes.createPolicy('default', {
+					// https://github.com/angular/angular/issues/31329 can't use Angular DomSanitizer here
+					createHTML: (s) => {
+						return DOMPurify.sanitize(
+							s
+								// hack so chrome won't complain about inline styles (mainly from svg icons)
+								.replace('<style />', ''),
+							{
+								USE_PROFILES: { html: true, svg: true },
+							},
+						);
+					},
+				});
+			},
+			multi: true,
+		},
+		{
+			provide: APP_INITIALIZER,
+			useFactory: (http: HttpClient) => async () => {
+				const config = await firstValueFrom(http.get('./assets/config.json'));
+				Object.assign(environment, { ...config, ...environment });
+			},
+			deps: [HttpClient],
+			multi: true,
+		},
+		{ provide: NZ_I18N, useValue: de_DE },
 	],
 	bootstrap: [AppComponent],
 })
