@@ -1,12 +1,11 @@
+import { nxE2EPreset } from '@nx/playwright/preset';
 import type { PlaywrightTestConfig } from '@playwright/test';
 import { devices } from '@playwright/test';
-
-import { baseConfig } from '../../playwright.config.base';
 
 const baseURL = process.env.E2E_BASE_URL || 'http://localhost:4200/';
 
 const config: PlaywrightTestConfig = {
-	...baseConfig,
+	...nxE2EPreset(__filename, { testDir: './src' }),
 	projects: [
 		{ name: 'setup', testMatch: /.*\.setup\.ts/ },
 		{
@@ -21,6 +20,7 @@ const config: PlaywrightTestConfig = {
 		baseURL,
 		screenshot: 'only-on-failure',
 		video: 'retry-with-video',
+		trace: 'on-first-retry',
 	},
 };
 
