@@ -4,7 +4,7 @@ import { TraceComponent } from '@kordis/spa/observability';
 
 // placeholder until we have a feature structure
 @Component({
-	selector: 'kordis-root',
+	selector: 'kordis-protected',
 	template: ` ganz geheim `,
 	changeDetection: ChangeDetectionStrategy.OnPush,
 })
