@@ -1,8 +1,17 @@
-import { Environment } from './environment.model';
+import { AuthConfig } from 'angular-oauth2-oidc';
 
-export const environment: Environment = {
+export const environment = {
 	production: false,
-	environmentName: 'Dev Local',
 	releaseVersion: '0.0.0-development',
-	apiUrl: 'http://localhost:3000',
+} as {
+	production: boolean;
+	releaseVersion: string;
+	// following properties are from the runtime dependent assets/config.json
+	environmentName: string;
+	apiUrl: string;
+	sentryKey?: string;
+	oauth?: {
+		config: AuthConfig;
+		discoveryDocumentUrl: string;
+	};
 };
